Workaround a require order problem in dm-migrations
This will be fixed in dm-migrations rather soon.
Currently, this gives dm-migrations to every
dm-rails user (as long as the line doesn't get
removed of course). In the future, all that will
be necessary is to have dm-migrations in the
Gemfile in case one wants to use it, and the
adapter supports it.